Display saving spinner when notes are updating
The new editor has support for showing a 'saving' message with spinner
while a note is being updated.  It also has both a html_content and
text content areas for future markdown editing support.  This updates
the workspace annotation editor to support both features, bringing it
on par feature wise with the embedded annotation editor